HFT-MakerBot/requirements.txt
python-binance==1.0.17
urllib3==1.26.18
requests==2.31.0
six==1.16.0
